FilePathCorrection
Homework file path worked correct. Had to redo the lab one as it was not sending you to the readme links. Believe I was missing part of the path name (lab/ before the labs since I made a subfolder for them). Added the correct path part and attempt number 2 at the table.