Corrected Code
Updated the code and realized that I did not have the words capitalized in the for loop so it was not actually pulling any information from the file. Rewrote the if/elif to represent the shape names with a capital letter and code worked on the first try after that.