Lab5
Worked through lab 5. Was having trouble with running the code but redownloaded the Garage.csv file and the campus.gdb file and saved them in my lab5 folder and then ran it from there. Think the issue was in the "geodatabase name" tab in the toolbox on ArcPro, I kept forgetting to add ".gdb" at the end of it cause as soon as I tried that with the newly downloaded files, it ran on the first try and produced the basis for the final screenshot I made (changing coloring and a few other visual things).